+++ src/ts/main.ts
@@ -12,6 +12,7 @@ export { PlayerUtils } from './utils/PlayerUtils';
 export { UIUtils } from './utils/UIUtils';
 export { BrowserUtils } from './utils/BrowserUtils';
 export { StorageUtils } from './utils/StorageUtils';
+export { UIPreferencesManager } from './utils/UIPreferencesManager';
 export { ErrorUtils } from './utils/ErrorUtils';
 export { ShadowDomManager } from './utils/ShadowDomManager';
 // Localization
@@ -138,6 +139,7 @@ export {
   DynamicSettingsPanelItem,
   DynamicSettingsPanelItemConfig,
 } from './components/settings/DynamicSettingsPanelItem';
+export { ToggleSettingsPanelItem, ToggleSettingsPanelItemConfig } from './components/settings/ToggleSettingsPanelItem';
 export { ReplayButton } from './components/buttons/ReplayButton';
 export { QuickSeekButton, QuickSeekButtonConfig } from './components/buttons/QuickSeekButton';
 export {
@@ -153,6 +155,7 @@ export { Icon } from './components/Icon';
 export { DismissClickOverlay, DismissClickOverlayConfig } from './components/overlays/DismissClickOverlay';
 export { EcoModeContainer } from './components/EcoModeContainer';
 export { EcoModeToggleButton } from './components/buttons/EcoModeToggleButton';
+export { PersistentPreferencesToggleButton } from './components/buttons/PersistentPreferencesToggleButton';
 export { TvNoiseCanvas } from './components/TvNoiseCanvas';
 export { SmallCenteredPlaybackToggleButton } from './components/buttons/SmallCenteredPlaybackToggleButton';
 export {